+++ codex/serializers/browser/page.py
@@ -12,7 +12,8 @@ from codex.serializers.browser.mixins import (
     BrowserAggregateSerializerMixin,
 )
 from codex.serializers.fields import TimestampField
-from codex.serializers.fields.browser import BreadcrumbsField, TopGroupField
+from codex.serializers.fields.browser import BreadcrumbsField
+from codex.serializers.fields.group import BrowseGroupField
 
 
 class BrowserCardSerializer(BrowserAggregateSerializerMixin, Serializer):
@@ -58,7 +59,7 @@ class BrowserPageSerializer(Serializer):
     title = BrowserTitleSerializer(read_only=True)
     zero_pad = IntegerField(read_only=True)
     libraries_exist = BooleanField(read_only=True)
-    model_group = TopGroupField(read_only=True)
+    model_group = BrowseGroupField(read_only=True)
     num_pages = IntegerField(read_only=True)
     groups = BrowserCardSerializer(allow_empty=True, read_only=True, many=True)
     books = BrowserCardSerializer(allow_empty=True, read_only=True, many=True)