+++ codex/serializers/fields/session.py
@@ -1,16 +1,11 @@
 """Session Custom Fields."""
 
-from rest_framework.fields import ChoiceField
-
 from codex.choices.browser import BROWSER_DEFAULTS
 from codex.choices.reader import READER_DEFAULTS
+from codex.serializers.fields.base import CodexChoiceField
 
-_SESSION_KEYS = (*READER_DEFAULTS.keys(), *BROWSER_DEFAULTS.keys(), "filters")
 
-
-class SessionKeyField(ChoiceField):
+class SessionKeyField(CodexChoiceField):
     """Session Key Field."""
 
-    def __init__(self, *args, **kwargs):
-        """Add choices."""
-        super().__init__(*args, choices=_SESSION_KEYS, **kwargs)
+    class_choices = (*READER_DEFAULTS.keys(), *BROWSER_DEFAULTS.keys(), "filters")