+++ codex/serializers/reader.py
@@ -1,25 +1,23 @@
 """Codex Reader Serializers."""
 
-from rest_framework.serializers import (
+from rest_framework.fields import (
     BooleanField,
     CharField,
     DecimalField,
+    DictField,
     IntegerField,
-    Serializer,
+    ListField,
 )
+from rest_framework.serializers import Serializer
+from typing_extensions import override
 
-from codex.serializers.browser.filters import BrowserSettingsFilterInputSerializer
-from codex.serializers.browser.settings import BrowserSettingsShowGroupFlagsSerializer
 from codex.serializers.fields import (
     FitToField,
     ReadingDirectionField,
     TimestampField,
 )
-from codex.serializers.fields.browser import (
-    BreadcrumbsField,
-    TopGroupField,
-)
-from codex.serializers.route import RouteSerializer, SimpleRouteSerializer
+from codex.serializers.fields.reader import ArcGroupField
+from codex.serializers.route import RouteSerializer
 
 
 class ReaderSettingsSerializer(Serializer):
@@ -44,24 +42,26 @@ class ReaderComicSerializer(Serializer):
     mtime = TimestampField(read_only=True)
 
 
-class ReaderArcSerializer(SimpleRouteSerializer):
+class ReaderArcInfoSerializer(Serializer):
     """Information about the current Arc."""
 
-    count = IntegerField(required=False, read_only=True)
-    index = IntegerField(required=False)
-    filters = BrowserSettingsFilterInputSerializer(required=False)
+    name = CharField(read_only=True)
     mtime = TimestampField(read_only=True)
-    name = CharField(required=False, read_only=True)
+
+
+class ReaderSelectedArcSerializer(Serializer):
+    """And arc key serializer."""
+
+    group = ArcGroupField(required=False)
+    ids = ListField(child=IntegerField(), required=False)
+    index = IntegerField(read_only=True, required=False)
+    count = IntegerField(read_only=True, required=False)
 
 
 class ReaderViewInputSerializer(Serializer):
     """Input for the reader serailizer."""
 
-    arc = ReaderArcSerializer(required=False)
-    breadcrumbs = BreadcrumbsField(required=False)
-    browser_arc = ReaderArcSerializer(required=False)
-    show = BrowserSettingsShowGroupFlagsSerializer(required=False)
-    top_group = TopGroupField(required=False)
+    arc = ReaderSelectedArcSerializer(required=False)
 
 
 class ReaderCurrentComicSerializer(ReaderComicSerializer):
@@ -78,11 +78,7 @@ class ReaderCurrentComicSerializer(ReaderComicSerializer):
         required=False,
     )
     issue_suffix = CharField(read_only=True, required=False)
-    issue_count = IntegerField(
-        read_only=True,
-        required=False,
-    )
-
+    issue_count = IntegerField(read_only=True, required=False)
     file_type = CharField(read_only=True, required=False)
     filename = CharField(read_only=True, required=False)
     name = CharField(read_only=True, required=False)
@@ -92,15 +88,37 @@ class ReaderBooksSerializer(Serializer):
     """All comics relevant to the reader."""
 
     current = ReaderCurrentComicSerializer(read_only=True)
-    prev_book = ReaderComicSerializer(read_only=True, required=False)
-    next_book = ReaderComicSerializer(read_only=True, required=False)
+    prev = ReaderComicSerializer(read_only=True, required=False)
+    next = ReaderComicSerializer(read_only=True, required=False)
+
+
+class ArcsIdsField(DictField):
+    """Arcs Ids level."""
+
+    @override
+    def to_representation(self, value):
+        """Serialize the ids to a string."""
+        string_keyed_map = {}
+        for ids, arc_info in value.items():
+            string_key = ",".join(str(pk) for pk in ids)
+            string_keyed_map[string_key] = arc_info
+
+        return super().to_representation(string_keyed_map)
+
+    child = ReaderArcInfoSerializer(read_only=True)
+
+
+class ArcsField(DictField):
+    """Arcs Field."""
+
+    child = ArcsIdsField(read_only=True)
 
 
 class ReaderComicsSerializer(Serializer):
     """Books and arcs."""
 
+    arcs = ArcsField(read_only=True)
+    arc = ReaderSelectedArcSerializer(read_only=True)
     books = ReaderBooksSerializer(read_only=True)
-    arcs = ReaderArcSerializer(many=True, read_only=True)
-    arc = ReaderArcSerializer(read_only=True)
     close_route = RouteSerializer(read_only=True)
     mtime = TimestampField(read_only=True)